moved firebase to frontend folder
