+++ frontend/src/pages/index.tsx
@@ -15,8 +15,8 @@ export default function Home() {
         <link rel="icon" href="/favicon.ico" />
       </Head>
       <main>
-        {/* <NavBar />
-        <BookApp /> */}
+        <NavBar />
+        <BookApp />
       </main>
     </>
   )
